后台单词本上传功能完善

+++ database/migrations/2017_08_12_213035_create_wordbooks_table.php
@@ -16,7 +16,7 @@ class CreateWordbooksTable extends Migration
         Schema::create('wordbooks', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->tinyInteger('sort')->default(0);
+            $table->integer('sort')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
